+++ x/authz/keeper/msg_server_test.go
@@ -369,3 +369,7 @@ func (suite *TestSuite) TestExec() {
 		})
 	}
 }
+
+func (suite *TestSuite) TestPruneExpiredGrants() {
+
+}